+++ CHANGELOG.rst
@@ -56,7 +56,7 @@ Changes:
 ^^^^^^^^
 
 - Added support for ``pem.OpenSSHPrivateKey`` (``OPENSSH PRIVATE KEY``).
-  OpenSSH added a new ``BEGIN`` label when it switched to a `proprietary key encoding <https://coolaj86.com/articles/the-openssh-private-key-format/>`_.
+  OpenSSH added a new ``BEGIN`` label when it switched to a proprietary key encoding.
   `#39 <https://github.com/hynek/pem/pull/39>`_